Add files via upload
This project is a career survey about data and has received 630 responses. I have used Power BI to create a report on average salaries, which is an important factor in deciding to pursue a career related to data. The goal of the project is to gather insights into the opinions and experiences of individuals in the data field and to analyze the data to identify trends and patterns. Additionally, the project aims to explore whether working with data is an accessible and feasible career choice for individuals who are interested in pursuing it.